==> notebooks/1_pickle_files.ipynb <==
**Dealing with a large dataset on your computer** is a common challenge for data analysts. If you need to process or filter the data you can set the `chunksize=` argument of pandas `read_csv()` method to loop through and work with manageable chunks of the data.

If for example, you wanted to work with a large data file (HCPS_data.csv) to pull just the rows where the `HCPS Code` is **99213**, you could read that file to a chunk of 100,000 rows at a time, filter each chunk to the rows with the specified code, save the filtered results of each chunk to a list and concatenate them all together at the end. The syntax would look something like this:

```
code_99213_rows =[]

for chunk in pd.read_csv('HCPS_data.csv', chunksize = 100000):
    code_99213_rows.append(chunk[chunk['HCPS Code'] == '99213']) 
               
code_99213_df = pd.concat(code_99213_rows, ignore_index=True)
```
======================================================================   

To shrink the size of a file so that it loads more quickly, converting a text file (CSV) to binary might make sense. In python, you can work with data to minimize its footprint and then store the resulting object (dataframe) as a [pickle](https://docs.python.org/3/library/pickle.html) file.

import pandas as pd
import pickle

%%time
may = pd.read_csv('../data/may.csv')
may.head()

### Now try to reduce the size of the file
- objects require the most space


may.info()

#### convert the company name to an integer 
- find the unique company names
- assign each company an integer (you can use a dictionary for this step)
- update the `companyname` column to store the integer id for each company

may.companyname.unique()

company_dict = {'Bird':0, 'Lyft': 1, 'Gotcha': 2, 'Lime': 3, 'Spin': 4, 'Jump': 5, 'Bolt': 6}

may.companyname = may.companyname.replace(company_dict)

#### next convert `pubdatetime` to a datetime 

may.pubdatetime = pd.to_datetime(may.pubdatetime)
may.head(2)

#### Next remove unneeded data
#### keep just the scooters

may.sumdgroup.unique()

may_scooters = may.loc[may.sumdgroup.isin(['scooter', 'Scooter'])]

#### keep just the columns we want to work with

may_scooters = may_scooters[['pubdatetime', 'latitude', 'longitude', 'sumdid', 'chargelevel', 'companyname']]

#### check `.info()` again

may_scooters.info()

#### The only object datatype remaining is sumdid (an alphanumeric unique identifier)
- time to pickle

may_scooters.to_pickle("../data/may.pkl")

%%time
may_test = pd.read_pickle("../data/may.pkl")

June

june = pd.read_csv('../data/june.csv')

june.companyname = june.companyname.replace(company_dict)

june.pubdatetime = pd.to_datetime(june.pubdatetime)
june.head(2)

june_scooters = june.loc[june.sumdgroup.isin(['scooter', 'Scooter'])]

june_scooters = june_scooters[['pubdatetime', 'latitude', 'longitude', 'sumdid', 'chargelevel', 'companyname']]

june_scooters.to_pickle("../data/june.pkl")

%%time
june_test = pd.read_pickle("../data/june.pkl")

July Pickle

july = pd.read_csv('../data/july.csv')

july.companyname = july.companyname.replace(company_dict)

july.pubdatetime = pd.to_datetime(july.pubdatetime)
july.head(2)

july_scooters = july.loc[july.sumdgroup.isin(['scooter', 'Scooter'])]

july_scooters = july_scooters[['pubdatetime', 'latitude', 'longitude', 'sumdid', 'chargelevel', 'companyname']]

july_scooters.to_pickle("../data/july.pkl")

%%time
july_test = pd.read_pickle("../data/july.pkl")
