==> notebooks/2_EDA_GEOSPATIALv2.ipynb <==
import pandas as pd
import numpy as np
from shapely.geometry import Point
import geopandas as gpd
import matplotlib.pyplot as plt
import folium
from folium.plugins import MarkerCluster
from folium.plugins import FastMarkerCluster

may = pd.read_pickle('../data/may.pkl')
june = pd.read_pickle('../data/june.pkl')
july = pd.read_pickle('../data/july.pkl')

may.head(10)

may.tail(10)

company0 = may.loc[may['sumdid'] == 'PoweredLIRL1']

company0

company3 = may.loc[may['sumdid'] == 'PoweredHICJ5XYHBAW5I']

company3

NUMBER OF UNIQUE SCOOTERS - MAY

unique_may = may.nunique()

unique_may

unique_june = june.nunique()

unique_june

june_unique_company = june.groupby("companyname")

june_unique_company = june_unique_company.agg({"sumdid": "nunique"})

june_unique_company

july_unique_company = july.groupby("companyname")

july_unique_company = july_unique_company.agg({"sumdid": "nunique"})

july_unique_company

unique_july = july.nunique()

unique_july

may.info

may.isnull().sum()

may.companyname.value_counts()

may_grouped2 = may.groupby("companyname")

may_grouped2 = may_grouped2.agg({"sumdid": "nunique"})

may_grouped2

june_grouped2 = june.groupby("companyname")

june_grouped2 = june_grouped2.agg({"sumdid": "nunique"})

june_grouped2

june_grouped2 = june.groupby("companyname")

june_grouped2 = june_grouped2.agg({"sumdid": "nunique"})

june_grouped2

may_sorted = may.sort_values(by=['sumdid', 'pubdatetime'])

may_sorted.head()

may.loc['2019-05-01 1:00:00':'2019-05-01 23:59:42.000']

All pings in 1 day

may_1 = may[(may['pubdatetime'] > '2019-05-01 0:00:00') & (may['pubdatetime'] < '2019-05-01 23:59:42.000')]

may_1

may_1_1scooter = may_1.loc[may['sumdid'] == 'PoweredLIRL1']

may_1_1scooter

this made my hours column

may['hours'] = pd.to_datetime(may['pubdatetime']).dt.hour


lets me filter by a range of those hou

may.loc[(may['hours'] > 6) & (may['hours'] < 10)]

let me search by a specific date from my dates column

may.loc[may["dates"]== datetime.date(2019, 5, 1)]
