==> notebooks/3_exploration_june.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import datetime
import numpy as np

june = pd.read_pickle('../data/june.pkl')
junetrip = pd.read_csv('../data/june_trip.csv')

june.head()

june.tail()

june.info()

#note: there are 8 million more rows of data in June than there are in May.  The month is actually shorter than may, so it's not that ... Maybe more logging per scooter?  maybe more scooters in total?

june['sumdid'].value_counts() 

nope - fewer scooters.  there were 8303 in may, there's 7822 in june.
I still see the top number of entries is about the same, and the bottom is still 1

june['companyname'].value_counts()

interesting - in may, company 3 had the most entries.  in june, it's company 5.  they stepped up their data collection.  do they have more scooters too?  Even though the overall number is smaller?

june_company_group = june.groupby("companyname")

june_scooter_company = june_company_group.agg({"sumdid": "nunique"})
print(june_scooter_company)

company 0 dropped 154 scooters;
company 1 dropped 224 scooters;
company 2 same;
company 3 dropped 176 scooters;
company 4 gained 26 scooters;
company 5 gained 60 scooters;
company 6 dropped 13 scooters

june.isnull().sum()

june_nullcharge_company= june.chargelevel.isnull().groupby(june.companyname).sum()
print(june_nullcharge_company)

i was careful to make sure that the company name/numbers were the same across all 3 months.  May all nulls were in company 4.   june, all nulls are in company 6.  what changed?

just for kicks, let's see how many single value entries there are in june ... 

june.groupby("sumdid").filter(lambda x: len(x) == 1)

just 8 this time.  off hand, looks like different scooters  similar distribution across companies, dates, locations ... 

repeating what i learned playing in may's data to june

june['dates'] = pd.to_datetime(june['pubdatetime']).dt.date
june['times'] = pd.to_datetime(june['pubdatetime']).dt.time
june['hours'] = pd.to_datetime(june['pubdatetime']).dt.hour

june.head()

#june.loc[june["dates"]== datetime.date(2019, 6, 1)]
#to filter my df by a date

june.loc[june["hours"]==(7)]
#to look at a specific hour window

june.loc[(june['hours'] > 6) & (june['hours'] < 10)]
#a range of time - the morning commute

#NOW TO LOOK AT THE TRIPS DF

junetrip.head()

junetrip.tail()

junetrip.info()

junetrip.companyName.unique()

company_dict = {'Bird':0, 'Lyft': 1, 'Gotcha': 2, 'Lime': 3, 'SPIN': 4, 'JUMP': 5, 'Bolt Mobility': 6}
junetrip.companyName = junetrip.companyName.replace(company_dict)

junetrip.columns

junetrip = junetrip.rename(columns={'pubTimeStamp':'pubdatetime', 'companyName':'companyname', 'tripRecordNum':'tripnum', 'sumdID':'sumdid','tripDuration':'tripduration', 'tripDistance':'tripdistance', 'startDate':'startdate', 'startTime':'starttime', 'endDate':'enddate', 'endTime':'endtime', 'startLatitude':'startlat', 'startLongitude':'startlong', 'endLatitude':'endlat', 'endLongitude':'endlong', 'tripRoute':'triproute'})
junetrip.head()

junetrip.pubdatetime = pd.to_datetime(junetrip.pubdatetime)
junetrip['dates'] = pd.to_datetime(junetrip['pubdatetime']).dt.date
junetrip['times'] = pd.to_datetime(junetrip['pubdatetime']).dt.time
junetrip['hours'] = pd.to_datetime(junetrip['pubdatetime']).dt.hour
junetrip.head()

REPEATING WORK FROM MAY INTO JUNE NOTEBOOK TO GATHER ABANDONED SCOOTER DATA

junesorted = june.sort_values(['sumdid', 'pubdatetime'])

junesorted['latmovement'] = (junesorted.latitude.shift() - junesorted.latitude)

june_stationaryscooter= junesorted[junesorted['latmovement']==0.0]

june_stationaryscooter['lonmovement'] = (june_stationaryscooter.longitude.shift() - june_stationaryscooter.longitude)

june_stationaryscooter = june_stationaryscooter[june_stationaryscooter['lonmovement']==0.0]

june_stationaryscooter.head()

june_stationaryscooter.info()

may_stationaryscooter.head()

june_stationaryscooter['timeshift']= (june_stationaryscooter.pubdatetime - june_stationaryscooter.pubdatetime.shift())

june_stationaryscooter['timeshift'] = june_stationaryscooter['timeshift']/np.timedelta64(1, 's')

def duration(row): 
    if 240 <= row ['timeshift'] <= 360:
        return 5
    else:
        return 0

june_stationaryscooter['duration'] = june_stationaryscooter.apply(duration, axis = 1)

june_stationaryscooter.head(10)

june_stationaryscooter.to_pickle("../data/june_stationaryscooter.pkl")
