==> notebooks/4_eda.ipynb <==
import pandas as pd
import pickle

Now, I want to do some basic exploration.

may = pd.read_pickle("../data/may.pkl")
may.head()

may.tail()

may.info()

may.isnull().sum()

may.describe()

grouped_may = may.groupby('companyname')
grouped_may = grouped_may.agg({'sumdid': 'nunique'})
print(grouped_may)

Now, I want to separate the data by company to look at individual characteristics.

company0_may = may.loc[may['companyname'] == 0]
company0_may.head()

company1_may = may.loc[may['companyname'] == 1]
company1_may.head()

company2_may = may.loc[may['companyname'] == 2]
company2_may.head()

company3_may = may.loc[may['companyname'] == 3]
company3_may.head()

company4_may = may.loc[may['companyname'] == 4]
company4_may.head()

company5_may = may.loc[may['companyname'] == 5]
company5_may.head()

company6_may = may.loc[may['companyname'] == 6]
company6_may.head()

Next, I want to look at unique counts for different scooters just to get a feel for the upper ranges of usage. 

print(company0_may['sumdid'].value_counts())

print(company1_may['sumdid'].value_counts())

print(company2_may['sumdid'].value_counts())

print(company3_may['sumdid'].value_counts())

print(company4_may['sumdid'].value_counts())

print(company5_may['sumdid'].value_counts())

print(company6_may['sumdid'].value_counts())

Just to get an idea of how often a scooter is pinged when its active, I'm going to look at one of the most active scooters for company 0.

scooter = company0_may.loc[company0_may['sumdid'] == 'PoweredLIRL1']
print(scooter.tail(25))

june = pd.read_pickle("../data/june.pkl")
june.head()

june.tail()

june.info()

june.isnull().sum()

june.describe()

grouped_june = june.groupby('companyname')
grouped_june = grouped_june.agg({'sumdid': 'nunique'})
print(grouped_june)

july = pd.read_pickle("../data/july.pkl")
july.head()

july.tail()

july.info()

july.isnull().sum()

grouped_july = july.groupby('companyname')
grouped_july = grouped_july.agg({'sumdid': 'nunique'})
print(grouped_july)
