==> notebooks/4_exploration_july.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import datetime
import numpy as np

july = pd.read_pickle('../data/july.pkl')
julytrip = pd.read_csv('../data/july_trip.csv')

july.head()

july.tail()

july.info()

july is 3 million rows smaller than june, even though the month is longer and should have more data as a result ... 

july['sumdid'].value_counts() 

there are even fewer scooters in july than in june - down another 1035.  more tourists in the summer, but fewer students?

july['companyname'].value_counts()

company 5 is at the top again.  
company 3 slid down another spot in data entry.  
company 0 moved up a spot

july_company_group = july.groupby("companyname")
july_scooter_company = july_company_group.agg({"sumdid": "nunique"})
print(july_scooter_company)

company 0 dropped another 327;
company 1 dropped another 126; 
company 2 gained 1 scooter;
company 3 dropped another 553;
company 4 dropped 143;
company 5 gained 170 scooters;
company 6 dropped another 57 scooters

july.isnull().sum()

july_nullcharge_company= july.chargelevel.isnull().groupby(july.companyname).sum()
print(july_nullcharge_company)

company 6 has all the nulls again

july.groupby("sumdid").filter(lambda x: len(x) == 1)

11 scooters this time with just 1 entry - companies 3 and 4 the biggest offenders yet again

repeating what i learned in may and applying to july

july['dates'] = pd.to_datetime(july['pubdatetime']).dt.date
july['times'] = pd.to_datetime(july['pubdatetime']).dt.time
july['hours'] = pd.to_datetime(july['pubdatetime']).dt.hour

july.head()

#july.loc[july["dates"]== datetime.date(2019, 7, 1)]
#to filter my df by a date

#july.loc[july["hours"]==(7)]
#to look at a specific hour window

#july.loc[(july['hours'] > 6) & (july['hours'] < 10)]
#a range of time - the morning commute

#NOW TO LOOK AT THE TRIPS DF

julytrip.head()

julytrip.tail()

julytrip.info()

julytrip.companyName.unique()

company_dict = {'Bird':0, 'Lyft': 1, 'Gotcha': 2, 'Lime': 3, 'SPIN': 4, 'JUMP': 5, 'Bolt Mobility': 6}
julytrip.companyName = julytrip.companyName.replace(company_dict)

julytrip.columns

julytrip = julytrip.rename(columns={'pubTimeStamp':'pubdatetime', 'companyName':'companyname', 'tripRecordNum':'tripnum', 'sumdID':'sumdid','tripDuration':'tripduration', 'tripDistance':'tripdistance', 'startDate':'startdate', 'startTime':'starttime', 'endDate':'enddate', 'endTime':'endtime', 'startLatitude':'startlat', 'startLongitude':'startlong', 'endLatitude':'endlat', 'endLongitude':'endlong', 'tripRoute':'triproute'})
julytrip.head()

julytrip.pubdatetime = pd.to_datetime(julytrip.pubdatetime)
julytrip['dates'] = pd.to_datetime(julytrip['pubdatetime']).dt.date
julytrip['times'] = pd.to_datetime(julytrip['pubdatetime']).dt.time
julytrip['hours'] = pd.to_datetime(julytrip['pubdatetime']).dt.hour
julytrip.head()

REPEATING WORK FROM MAY INTO JULY NOTEBOOK TO FIND ABANDONED SCOOTER DATA

julysorted = july.sort_values(['sumdid', 'pubdatetime'])

julysorted['latmovement'] = (julysorted.latitude.shift() - julysorted.latitude)

july_stationaryscooter= julysorted[julysorted['latmovement']==0.0]

july_stationaryscooter['lonmovement'] = (july_stationaryscooter.longitude.shift() - july_stationaryscooter.longitude)

july_stationaryscooter = july_stationaryscooter[july_stationaryscooter['lonmovement']==0.0]

july_stationaryscooter.head()

july_stationaryscooter['timeshift']= (july_stationaryscooter.pubdatetime - july_stationaryscooter.pubdatetime.shift())

july_stationaryscooter['timeshift'] = july_stationaryscooter['timeshift']/np.timedelta64(1, 's')

def duration(row): 
    if 240 <= row ['timeshift'] <= 360:
        return 5
    else:
        return 0

july_stationaryscooter['duration'] = july_stationaryscooter.apply(duration, axis = 1)

july_stationaryscooter.head(10)

#july_stationaryscooter.to_pickle("../data/july_stationaryscooter.pkl")
